Add dep to scala3-binary-decoder
