--- modules/decoder/src/main/scala/ch/epfl/scala/debugadapter/internal/binary/BinaryClassLoader.scala
@@ -1,4 +1,0 @@
-package ch.epfl.scala.debugadapter.internal.binary
-
-trait BinaryClassLoader:
-  def loadClass(name: String): ClassType
