--- modules/decoder/src/main/scala/ch/epfl/scala/debugadapter/internal/binary/Field.scala
@@ -1,5 +1,0 @@
-package ch.epfl.scala.debugadapter.internal.binary
-
-trait Field extends Symbol:
-  def declaringClass: ClassType
-  def `type`: Type
