--- modules/decoder/src/main/scala/ch/epfl/scala/debugadapter/internal/binary/SignedName.scala
@@ -1,3 +1,0 @@
-package ch.epfl.scala.debugadapter.internal.binary
-
-final case class SignedName(name: String, signature: String)
