--- modules/decoder/src/main/scala/ch/epfl/scala/debugadapter/internal/binary/Symbol.scala
@@ -1,10 +1,0 @@
-package ch.epfl.scala.debugadapter.internal.binary
-
-import ch.epfl.scala.debugadapter.internal.stacktrace.NameTransformer
-
-trait Symbol:
-  def name: String
-  def sourceLines: Option[SourceLines]
-
-  protected def showSpan: String =
-    sourceLines.map(_.showSpan).getOrElse("")
