--- modules/decoder/src/main/scala/ch/epfl/scala/debugadapter/internal/binary/Type.scala
@@ -1,3 +1,0 @@
-package ch.epfl.scala.debugadapter.internal.binary
-
-trait Type extends Symbol
