--- modules/decoder/src/main/scala/ch/epfl/scala/debugadapter/internal/javareflect/JavaReflectParameter.scala
@@ -1,14 +1,0 @@
-package ch.epfl.scala.debugadapter.internal.javareflect
-
-import ch.epfl.scala.debugadapter.internal.binary
-
-import java.lang.reflect.Parameter
-
-class JavaReflectParameter(parameter: Parameter, loader: JavaReflectLoader) extends binary.Parameter:
-
-  override def name: String = parameter.getName
-  override def sourceLines: Option[binary.SourceLines] = None
-
-  override def `type`: binary.Type = loader.loadClass(parameter.getType)
-
-  override def toString: String = parameter.toString
