+++ modules/decoder/src/main/scala/ch/epfl/scala/debugadapter/internal/jdi/JdiClassLoader.scala
@@ -1,15 +1,13 @@
 package ch.epfl.scala.debugadapter.internal.jdi
 
-import ch.epfl.scala.debugadapter.internal.binary
-import ch.epfl.scala.debugadapter.internal.binary.ClassType
-
+import ch.epfl.scala.decoder.binary
 import java.util as ju
 import scala.jdk.CollectionConverters.*
 
 class JdiClassLoader(obj: Any)
     extends binary.BinaryClassLoader
     with JavaReflection(obj, "com.sun.jdi.ClassLoaderReference"):
-  override def loadClass(name: String): ClassType =
+  override def loadClass(name: String): binary.ClassType =
     visibleClasses.find(_.name == name).get
 
   private def visibleClasses: Seq[JdiReferenceType] =
