+++ modules/decoder/src/main/scala/ch/epfl/scala/debugadapter/internal/jdi/JdiLocalVariable.scala
@@ -1,6 +1,6 @@
 package ch.epfl.scala.debugadapter.internal.jdi
 
-import ch.epfl.scala.debugadapter.internal.binary.*
+import ch.epfl.scala.decoder.binary.*
 
 class JdiLocalVariable(obj: Any) extends JavaReflection(obj, "com.sun.jdi.LocalVariable") with Parameter:
   override def name: String = invokeMethod("name")
