+++ modules/decoder/src/main/scala/ch/epfl/scala/debugadapter/internal/jdi/JdiType.scala
@@ -1,5 +1,5 @@
 package ch.epfl.scala.debugadapter.internal.jdi
-import ch.epfl.scala.debugadapter.internal.binary.*
+import ch.epfl.scala.decoder.binary.*
 
 class JdiType(obj: Any, className: String = "com.sun.jdi.Type") extends JavaReflection(obj, className) with Type:
   override def name: String = invokeMethod("name")
