--- modules/decoder/src/main/scala/ch/epfl/scala/debugadapter/internal/stacktrace/DecodedMethodBridge.scala
@@ -1,5 +1,0 @@
-package ch.epfl.scala.debugadapter.internal.stacktrace
-
-class DecodedMethodBridge(method: DecodedMethod, formatter: StackTraceFormatter):
-  def format: String = formatter.format(method)
-  def isGenerated: Boolean = method.isGenerated
