--- modules/decoder/src/test/resources/scala3-artifacts-231121-failed.csv
@@ -1,146 +1,0 @@
-"group_id","artifact_id","version"
-"io.github.palanga","parana-core-local_3","0.7.0"
-"io.github.palanga","parana-journal-common_3","0.7.0"
-"dev.atedeg.mdm","client-orders_3","4.0.0"
-"dev.atedeg.mdm","production-planning_3","4.0.0"
-"dev.atedeg.mdm","restocking_3","4.0.0"
-"dev.atedeg.mdm","milk-planning_3","4.0.0"
-"dev.atedeg.mdm","pricing_3","4.0.0"
-"com.github.vagmcs","optimus-solver-gurobi_3","3.4.3"
-"com.github.vagmcs","optimus-solver-mosek_3","3.4.3"
-"com.github.dmytromitin","scala3-compiler-patched-assembly_3","3.2.1"
-"com.thesamet.scalapb","protobuf-runtime-scala_3","0.8.14"
-"dev.optics","monocle-core_3","3.2.0"
-"cn.thinkingdata.android","ta-agp-v7_3","2.1.0"
-"com.typesafe.akka","akka-bom_3","2.8.5"
-"org.scalafx","scalafx-extras_3","0.8.0"
-"net.sf.ij-plugins","scala-console_3","1.8.0"
-"com.lightbend.akka","akka-projection-grpc_3","1.4.2"
-"com.phylage","refuel-json_3","2.1.1"
-"sh.almond","toree-hooks_3","0.13.14"
-"org.apache.pekko","pekko-cluster-metrics_3","1.0.1"
-"org.apache.pekko","pekko-bom_3","1.0.1"
-"org.apache.pekko","pekko-cluster-sharding-typed_3","1.0.1"
-"com.lightbend.akka.grpc","akka-grpc-runtime_3","2.3.4"
-"io.github.scala-ts","scala-ts-python_3","0.5.17"
-"uk.gov.nationalarchives","fs2_3","0.0.1"
-"com.didiglobal.booster","booster-android-gradle-v3_3","4.16.3"
-"org.scalatra","scalatra-cache-javax_3","3.0.0"
-"fr.hammons","slinc-j17_3","0.6.0"
-"org.scalatra","scalatra-common-jakarta_3","3.0.0"
-"org.scalatra","scalatra-forms-jakarta_3","3.0.0"
-"org.scalatra","scalatra-forms-javax_3","3.0.0"
-"org.scalatra","scalatra-jakarta_3","3.0.0"
-"org.scalatra","scalatra-javax_3","3.0.0"
-"org.scalatra","scalatra-json-javax_3","3.0.0"
-"org.scalatra","scalatra-json-jakarta_3","3.0.0"
-"io.github.aliyun-sls","android-gradle-v3_3","0.2.0"
-"io.github.aliyun-sls","android-gradle-v7_3","0.2.0"
-"org.apache.pekko","pekko-http-bom_3","1.0.0"
-"org.apache.pekko","pekko-http-testkit_3","1.0.0"
-"org.virtuslab","besom-core_3","0.1.0"
-"org.apache.pekko","pekko-connectors-jms_3","1.0.1"
-"io.github.ahjohannessen","sec-fs2-client_3","0.40.0"
-"io.shiftleft","js2cpg_3","0.2.246"
-"com.github.rssh","cps-async-connect-cats-effect-loom_3","0.9.19"
-"com.github.rssh","cps-async-connect-zio2-loom_3","0.9.19"
-"org.automorph","automorph-core_3","0.2.3"
-"io.kamon","kamon-apm-reporter_3","2.6.6"
-"io.kamon","kamon-datadog_3","2.6.6"
-"io.kamon","kamon-testkit_3","2.6.6"
-"com.typesafe.play","play_3","2.9.0"
-"com.typesafe.play","play-bom_3","2.9.0"
-"org.playframework","play-bom_3","3.0.0"
-"io.github.vigoo","zio-http-rust_3","0.0.14"
-"org.scalaz","scalaz_3","7.3.8"
-"de.tu-darmstadt.stg","rescala_3","0.34.0"
-"org.wvlet.airframe","airspec_3","23.11.3"
-"org.wvlet.airframe","airframe-http-grpc_3","23.11.3"
-"io.github.quafadas","vecxt_3","0.0.8"
-"io.github.philemone","hobogrid_3","0.0.1"
-"io.github.philemone","taminar_3","0.0.1"
-"com.softwaremill.sttp.tapir","tapir-nima-server_3","1.9.1"
-"com.softwaremill.sttp.tapir","tapir-core_3","1.9.1"
-"com.softwaremill.sttp.tapir","tapir-zio1_3","1.9.1"
-"com.softwaremill.sttp.tapir","tapir-zio_3","1.9.1"
-"io.joern","c2cpg_3","2.0.161"
-"io.joern","console_3","2.0.161"
-"io.joern","javasrc2cpg_3","2.0.161"
-"io.joern","kotlin2cpg_3","2.0.161"
-"io.joern","querydb_3","2.0.161"
-"io.joern","joern-cli_3","2.0.161"
-"io.joern","x2cpg_3","2.0.161"
-"io.taig","scala-pygments-graalvm_3","0.0.1"
-"io.taig","scala-pygments-graalvm-python_3","0.0.15"
-"io.taig","scala-linguist-graalvm-ruby_3","0.0.18"
-"io.kevinlee","logger-f-sbt-logging_3","1.20.0"
-"com.whisk","docker-testkit-core-shaded_3","0.11.0"
-"com.github.aaronp","franz_3","0.3.0"
-"com.typesafe","ssl-config-core_3","0.6.1"
-"io.github.vincenzobaz","encoders_3","0.1.1"
-"org.eclipse.pass","pass-client-shaded-v2_3","0.1.0"
-"com.babylonhealth.lit","generator_3","0.14.17"
-"dev.atedeg.mdm","production_3","4.0.0"
-"dev.atedeg.mdm","stocking_3","4.0.0"
-"com.babylonhealth.lit","protoshim_3","0.14.17"
-"org.typelevel","shapeless3-deriving_3","3.3.0"
-"com.github.scalaprops","scalaprops-all_3","0.9.1"
-"net.sf.ij-plugins","ijp-color-ui_3","0.12.2"
-"io.github.cquiroz","scala-java-locales_3","1.5.1"
-"io.monix","monix-java_3","3.4.1"
-"com.sfxcode.sapphire","sapphire-data_3","2.0.5"
-"com.tersesystems.blindsight","blindsight-inspection_3","2.0.0"
-"dev.zio","zio-interop-cats-tests_3","23.0.3"
-"io.chrisdavenport","mules-http4s-scodec_3","0.4.0"
-"io.getkyo","kyo-core-opt1_3","0.4.0"
-"io.getkyo","kyo-core-opt2_3","0.4.0"
-"io.getkyo","kyo-core-opt3_3","0.4.0"
-"com.typesafe.akka","akka-actor-testkit-typed_3","2.8.5"
-"com.typesafe.akka","akka-cluster-metrics_3","2.8.5"
-"com.typesafe.akka","akka-cluster-sharding-typed_3","2.8.5"
-"org.typelevel","log4cats-core_3","2.6.0"
-"org.typelevel","log4cats-slf4j_3","2.6.0"
-"net.sf.ij-plugins","scala-console-plugins_3","1.8.0"
-"io.bullet","borer-core_3","1.12.0"
-"com.softwaremill.ox","core_3","0.0.14"
-"org.apache.pekko","pekko-actor-testkit-typed_3","1.0.1"
-"org.bitlap","rolls-csv_3","0.3.4"
-"com.lightbend.akka.grpc","akka-grpc-interop-tests_3","2.3.4"
-"io.github.scala-ts","scala-ts-idtlt_3","0.5.17"
-"com.hinadt.hicloud.android","sa-agp-v7_3","2.0.0"
-"uk.gov.nationalarchives","zio_3","0.0.1"
-"com.typesafe.akka","akka-http-bom_3","10.5.3"
-"com.typesafe.akka","akka-http-testkit_3","10.5.3"
-"org.scalatest","scalatest-core_3","3.2.17"
-"com.didiglobal.booster","booster-android-gradle-v7_3","4.16.3"
-"org.scalatra","scalatra-cache-jakarta_3","3.0.0"
-"org.scalatra","scalatra-auth-jakarta_3","3.0.0"
-"org.scalatra","scalatra-auth-javax_3","3.0.0"
-"org.scalatra","scalatra-common-javax_3","3.0.0"
-"fr.hammons","slinc-j19_3","0.6.0"
-"org.scalatra","scalatra-swagger-javax_3","3.0.0"
-"org.scalatra","scalatra-twirl-jakarta_3","3.0.0"
-"org.scalatra","scalatra-twirl-javax_3","3.0.0"
-"com.xebia","xef-scala_3","0.0.3"
-"io.github.arainko","ducktape_3","0.1.11"
-"org.scalamolecule","molecule-datalog-datomic_3","0.7.0"
-"io.ino","solrs_3","2.8.0"
-"com.coralogix","zio-k8s-client_3","2.1.0"
-"com.github.fd4s","fs2-kafka-vulcan_3","3.2.0"
-"com.github.fd4s","fs2-kafka-vulcan-testkit-munit_3","3.2.0"
-"com.github.rssh","dotty-cps-async-loom_3","0.9.19"
-"edu.gemini","giapi_3","0.1.0"
-"io.github.mercurievv.minuscles","conversions_3","0.1.1"
-"org.scalatest","scalatest-app_3","3.2.17"
-"org.apache.pekko","pekko-grpc-runtime_3","1.0.1"
-"com.sensorsdata.analytics.android","sa-agp-v7_3","4.0.1"
-"io.github.embeddedkafka","embedded-kafka-schema-registry_3","7.5.2"
-"io.github.embeddedkafka","embedded-kafka-schema-registry-streams_3","7.5.2"
-"com.logicovercode","core-adts_3","0.0.1"
-"com.logicovercode","core-commons_3","0.0.1"
-"com.logicovercode","app-core_3","0.0.3"
-"com.linecorp.armeria","armeria-scala_3","1.26.3"
-"org.beangle.webmvc","beangle-webmvc-support_3","0.9.18"
-"io.github.guymers","zoobie_3","0.14.8"
-"io.7mind.izumi","distage-testkit-scalatest_3","1.2.1"
-"io.taig","scala-linguist-graalvm_3","0.0.4"
