--- modules/decoder/src/test/resources/test-result-231121-total.txt
@@ -1,11 +1,0 @@
-total classes (568075): 
-  - success: 542765 (95.5446%)
-  - ambiguous: 228 (0.040135544%)
-  - not found: 15399 (2.7107337%)
-  - throwables: 9683 (1.7045285%)
-total methods (8721744): 
-  - success: 8434059 (96.701515%)
-  - ambiguous: 16132 (0.184963%)
-  - not found: 131552 (1.5083222%)
-  - throwables: 140001 (1.605195%)
-
